Added Validation of Date and Time to Reminders and 30 min earlier it will notify the person , login logo in navabar with drop down menu done.

+++ run.py
@@ -7,8 +7,8 @@ from reminder_app.models import Reminder,User
 def mailer():
     with scheduler.app.app_context():
         d = datetime.datetime.now().date()
-        t = datetime.datetime.now().time()
-        r = Reminder.query.filter(Reminder.date == d and Reminder.time<=t).all()
+        t = (datetime.datetime.now() + datetime.timedelta(minutes=30)).time()
+        r = Reminder.query.filter(Reminder.date == d , Reminder.time<=t).all()
         for _ in r:
             with mail.connect() as conn:
                 user = User.query.filter_by(id = _.user_id).first()
@@ -16,8 +16,9 @@ def mailer():
                 subject = _.title
                 msg = Message(sender=os.environ.get('email_address'),recipients=[user.email],body=message,subject=subject)
                 conn.send(msg)
-            db.session.delete(_)
-            db.session.commit()    
+            if _:
+                db.session.delete(_)
+                db.session.commit()    
 if __name__=='__main__':
     scheduler.add_job(id="mailer",func=mailer,trigger='interval',seconds=60)
     scheduler.start()
